LJM_IMonster02
미니언, 공허충, 대포미니언 스텟 추가 / Atk 구문 int -> float 수정 / 쓰러지는 판정 /데미지 받기 추가 / Program.cs 각각 몬스터 생성

+++ Interfaces/IMonster.cs
@@ -11,7 +11,7 @@ namespace OceanStory.Interfaces
         string Name { get; }
         int Hp { get; set; }
         int Level { get; }
-        int Atk { get; }
+        float Atk { get; }
         bool MosterDead { get; }
         void GetDamage (int damage);
     }
